- SEC-6: Wire BKZ cost models into security estimator
- Eliminate dead code across all workspace crates (zero warnings)
- Resolve all clippy warnings (zero lint issues)
- Harden u128 CRT overflow guards + bootstrap correctness contract
- Harden Garner reconstruct overflow detection (checked_mul)
- Add secure_192/256 bootstrap overflow integration tests
- Wire U256 into bootstrap modswitch for >128-bit CRT products
- Harden public API panic paths + fallible MobiusInt ops
- Harden CRT assertion + RNSCiphertext::validate()
- Fix secure_256 params (226→264-bit hybrid security)
- Redact Debug on polynomial types (prevent secret key leakage)
- Harden polynomial eval bounds + hide unvalidated deserialization
- Softmax edge cases, ZeroizeOnDrop, panic sanitization, CRT boundary tests

+++ crates/clockwork-core/src/key_lifecycle.rs
@@ -402,14 +402,14 @@ mod tests {
             let r = next_rng(&mut rng) % TEST_Q;
             let shares = KeySharePair::split(s, r, TEST_Q);
 
-            let bin = (shares.share_a() as u64 * num_bins / TEST_Q) as usize;
+            let bin = (shares.share_a() * num_bins / TEST_Q) as usize;
             let bin = bin.min(num_bins as usize - 1);
             bin_counts[bin] += 1;
         }
 
         // Simple χ² test: expect ~num_samples/num_bins per bin
         // Integer-only: scale by 1000 to avoid floating-point
-        let expected_x1000 = (num_samples as u64 * 1000) / num_bins as u64;
+        let expected_x1000 = (num_samples as u64 * 1000) / num_bins;
 
         let chi_sq_x1000: u64 = bin_counts
             .iter()
